feat: Add ability to edit cards during review
This commit introduces a new feature that allows users to edit the content of a card while they are reviewing it.

When viewing the answer to a card, the user can press 'e' to open a new set of floating windows, one for each field of the note. The user can then edit the content of the fields in these windows.

To save the changes, the user can press '<leader>s'. This will update the note in Anki and close the edit windows.

This feature improves the user experience by allowing for quick corrections and updates to cards without having to leave the review session.